\name{clusterplots}
\alias{clusterplots}
\title{
Cluster groups plotting function. 
}
\description{
Function that plots the time series clusters generated by either of the functions: "tseriesca", "tseriescm" or "tseriescq".
}
\usage{
clusterplots(L, data)
}
\arguments{
  \item{L}{output list from the functions: "tseriesca", "tseriescm" or "tseriescq".}
  \item{data}{Data frame with the time series information.}
}
\details{
See the examples in the documentation files of "tseriesca", "tseriescm" or "tseriescq" for an example of this function's usage.
}
\value{
The function returns the plots of the time series clusters directly. 
}
\author{
Martell-Juarez, D.A.
}