\name{diagplots}
\alias{diagplots}
\title{
Diagnostic plots function.
}
\description{
Function that produces the diagnostic plots to assess the convergence of the Markov Chains generated by either of the functions: "tseriesca", "tseriescm" or "tseriescq". 
}
\usage{
diagplots(L)
}
\details{
See the examples in the documentation files of "tseriesca", "tseriescm" or "tseriescq" for an example of this function's usage. 
}
\arguments{
  \item{L}{output list from the functions: "tseriesca", "tseriescm" or "tseriescq".}
}
\value{
The function returns three different kinds of plots to assess convergence of the generated Markov Chain: trace plots, histograms and ergodic mean plots. 
}
\author{
Martell-Juarez, D.A.
}
