\name{gdp}
\alias{gdp}
\docType{data}
\title{
GDP per person employed from 1990 to 2012
}
\description{
This data set contains the yearly GDP per person employed from 1990 to 2012 for 121 countries.
}
\usage{data(gdp)}
\format{
Matrix with 20 rows and 121 columns. 
}
\source{
http://data.worldbank.org/indicator/SL.GDP.PCAP.EM.KD
}
\keyword{datasets}
