\name{houses}
\alias{houses}
\docType{data}
\title{
House price statistics in Scotland from 2004 to 2014. 
}
\description{
This data set contains the average price of houses from the 1st quarter of 2004 to the 4th quarter of 2014 by the local authority areas of Scotland
}
\usage{data(houses)}
\format{
Matrix with 44 rows and 33 columns. 
}
\source{
http://www.ros.gov.uk/public/news/quarterly_statistics.html
}
\references{
http://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/
}
\keyword{datasets}
