\name{stocks}
\alias{stocks}
\docType{data}
\title{
Mexican stock exchange market prices
}
\description{
This data set contains the monthly adjusted closing prices of 58 shares of the mexican stock exchange market from September 2006 to August 2011. 
}
\usage{data(stocks)}
\format{
Data frame with 60 rows and 58 columns. 
}
\source{
http://www.dowjones.com/factiva/
}
\references{
This is the data set used by Nieto-Barajas, L.E. & Contreras-Cristan, A. (2014) as application for their paper. 
}
\keyword{datasets}
