\name{tseriescm.out}
\alias{tseriescm.out}
\docType{data}
\title{
Output of tseriescm function for the Mexican stock exchange market prices dataset
}
\description{
This object contains the output of the function tseriescm for the example described in its documentation file.
}
\usage{data(tseriescm.out)}
\details{
See function tseriescm
}
\examples{
data(tseriescm.out)
}
\keyword{datasets}
