\name{tseriescq.out}
\alias{tseriescq.out}
\docType{data}
\title{
Output of tseriesca function for the House price statistics in Scotland dataset
}
\description{
This object contains the output of the function tseriescq for the example described in its documentation file.
}
\usage{data(tseriescq.out)}
\details{
See function tseriescq
}
\examples{
data(tseriescq.out)
}
\keyword{datasets}