\name{BNPTSclust-package}
\alias{BNPTSclust-package}
\alias{BNPTSclust}
\docType{package}
\title{A Bayesian Nonparametric Algorithm for Time Series Clustering}
\description{
This package performs the algorithm for time series clustering described
in Nieto-Barajas and Contreras-Cristan (2014). The package contains 
functions to work with annual, monthly and quarterly time series data. 

The main functions to accomplish the above are:

1) tseriesca 

2) tseriescm

3) tseriescq 
}
\details{
\tabular{ll}{
Package: \tab BNPTSclust\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2015-04-19\cr
License: \tab GPL2, GPL3\cr
}
For a comprehensive guide on how to use the package, refer to the vignette
attached to the package.
}
\author{
Martell-Juarez, D.A. and Nieto-Barajas, L.E.

Maintainer: David Alejandro Martell Juarez <alex91599@gmail.com>

}
\references{
Nieto-Barajas, L.E. and Contreras-Cristan, A. (2014) A Bayesian Nonparametric Approach for Time Series Clustering. \emph{Bayesian Analysis} \bold{Vol. 9, No. 1} 147--170.
}

\keyword{ package }
