\name{tseriesca.out}
\alias{tseriesca.out}
\docType{data}
\title{
Output of tseriesca function for the GDP per person employed dataset
}
\description{
This object contains the output of the function tseriesca for the example described in its documentation file.
}
\usage{data(tseriesca.out)}
\details{
See function tseriesca
}
\examples{
data(tseriesca.out)
}
\keyword{datasets}
